requirements.txt
telethon>=1.30.0
python-dotenv>=1.0.0
tqdm>=4.0.0
requests>=2.0.0
